+++ modules/fetch-upickle/src/main/scala/io/laminext/fetch/ResponseDecodeFailed.scala
@@ -1,5 +1,0 @@
-package io.laminext.fetch
-
-import org.scalajs.dom.experimental.Response
-
-class ResponseDecodeFailed(val error: Throwable, val response: Response) extends Throwable